Add packaged robotic cell for Telemachus project
The packaged robotic cell for the Telemachus project was created using FANUC ROBOGUIDE Version 9 (Rev.ZH) and requires RoboGuide to open and unpack. The package includes all robot TP programs and compiled Karel programs. The source Karel files in ASCII format are available in the folder. This package provides the complete robotic cell for anyone interested in opening and exploring it. The project is open for further development and contributions from the community.