Add package robotic cell for Gutenberg project
The package robotic cell for the Gutenberg project was created using FANUC ROBOGUIDE Version 9 (Rev.ZH) and requires RoboGuide to open and unpack. The package includes all robot TP programs and compiled Karel program. The source Karel file in ASCII format is available in the folder. This package provides the complete robotic cell for anyone interested in opening and exploring it. The project is open for further development and contributions from the community.